requirements.txt
matplotlib==3.7.0
numpy==1.24.4
opencv_python==4.6.0.66
opencv_python_headless==4.7.0.72
pandas==1.3.5
Pillow==9.3.0
pytest==7.2.1
sahi==0.11.8
setuptools==64.0.2
transformers